rename(... RegistrationDto): registration 패키지 삭제
