--- src/main/java/roomescape/registration/domain/reservation/domain/Reservation.java
@@ -1,94 +1,0 @@
-package roomescape.registration.domain.reservation.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
-import java.time.LocalDate;
-import java.util.Objects;
-import roomescape.member.domain.Member;
-import roomescape.reservationtime.domain.ReservationTime;
-import roomescape.theme.domain.Theme;
-
-@Entity
-public class Reservation {
-
-    private static final int NULL_ID = 0;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(nullable = false)
-    private LocalDate date;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private ReservationTime reservationTime;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Theme theme;
-
-    // todo: 나중에 예약 / 방탈출 분리하자.
-    // 방탈출: 날짜, 시간, 테마 / 예약: 방탈출, 멤버
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Member member;
-
-    public Reservation() {
-    }
-
-    public Reservation(long id, LocalDate date, ReservationTime reservationTime, Theme theme, Member member) {
-        this.id = id;
-        this.date = date;
-        this.reservationTime = reservationTime;
-        this.theme = theme;
-        this.member = member;
-    }
-
-    public Reservation(LocalDate date, ReservationTime reservationTime, Theme theme, Member member) {
-        this(NULL_ID, date, reservationTime, theme, member);
-    }
-
-    public void setMember(Member member) {
-        this.member = member;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public ReservationTime getReservationTime() {
-        return reservationTime;
-    }
-
-    public Theme getTheme() {
-        return theme;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Reservation that = (Reservation) o;
-        return id == that.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
