--- src/main/java/roomescape/registration/domain/reservation/dto/ReservationRequest.java
@@ -1,10 +1,0 @@
-package roomescape.registration.domain.reservation.dto;
-
-import java.time.LocalDate;
-
-public record ReservationRequest(
-        LocalDate date,
-        long timeId,
-        long themeId
-) {
-}
