--- src/main/java/roomescape/registration/domain/reservation/dto/ReservationResponse.java
@@ -1,19 +1,0 @@
-package roomescape.registration.domain.reservation.dto;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import roomescape.registration.domain.reservation.domain.Reservation;
-
-public record ReservationResponse(
-        long id,
-        String memberName,
-        String themeName,
-        LocalDate date,
-        LocalTime startAt
-) {
-    public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(reservation.getId(), reservation.getMember().getName(),
-                reservation.getTheme().getName(), reservation.getDate(),
-                reservation.getReservationTime().getStartAt());
-    }
-}
