--- src/main/java/roomescape/registration/domain/reservation/dto/ReservationTimeAvailabilityResponse.java
@@ -1,11 +1,0 @@
-package roomescape.registration.domain.reservation.dto;
-
-import java.time.LocalTime;
-import roomescape.reservationtime.domain.ReservationTime;
-
-public record ReservationTimeAvailabilityResponse(long timeId, LocalTime startAt, boolean alreadyBooked) {
-
-    public static ReservationTimeAvailabilityResponse fromTime(ReservationTime time, boolean alreadyBooked) {
-        return new ReservationTimeAvailabilityResponse(time.getId(), time.getStartAt(), alreadyBooked);
-    }
-}
