--- src/main/java/roomescape/registration/domain/waiting/dto/WaitingRequest.java
@@ -1,10 +1,0 @@
-package roomescape.registration.domain.waiting.dto;
-
-import java.time.LocalDate;
-
-public record WaitingRequest(
-        LocalDate date,
-        long timeId,
-        long themeId
-) {
-}
