--- src/main/java/roomescape/registration/domain/waiting/service/WaitingService.java
@@ -1,114 +1,0 @@
-package roomescape.registration.domain.waiting.service;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import roomescape.exception.RoomEscapeException;
-import roomescape.exception.model.WaitingExceptionCode;
-import roomescape.member.domain.Member;
-import roomescape.member.repository.MemberRepository;
-import roomescape.registration.domain.reservation.domain.Reservation;
-import roomescape.registration.domain.reservation.repository.ReservationRepository;
-import roomescape.registration.domain.waiting.domain.Waiting;
-import roomescape.registration.domain.waiting.domain.WaitingWithRank;
-import roomescape.registration.domain.waiting.dto.WaitingDto;
-import roomescape.registration.domain.waiting.dto.WaitingRequest;
-import roomescape.registration.domain.waiting.dto.WaitingResponse;
-import roomescape.registration.domain.waiting.repository.WaitingRepository;
-
-@Transactional
-@Service
-public class WaitingService {
-
-    private final WaitingRepository waitingRepository;
-    private final ReservationRepository reservationRepository;
-    private final MemberRepository memberRepository;
-
-    public WaitingService(WaitingRepository waitingRepository, ReservationRepository reservationRepository,
-                          MemberRepository memberRepository) {
-        this.waitingRepository = waitingRepository;
-        this.reservationRepository = reservationRepository;
-        this.memberRepository = memberRepository;
-    }
-
-    public WaitingResponse addWaiting(WaitingRequest waitingRequest, long memberId) {
-        validateAlreadyReservation(waitingRequest, memberId);
-        validateAlreadyWaiting(waitingRequest, memberId);
-
-        Reservation reservation = reservationRepository.findReservationByDateAndThemeIdAndReservationTimeId(
-                waitingRequest.date(),
-                waitingRequest.themeId(),
-                waitingRequest.timeId()
-        );
-        Member member = memberRepository.findMemberById(memberId)
-                .orElseThrow(() -> new RoomEscapeException(WaitingExceptionCode.MEMBER_INFO_IS_NULL_EXCEPTION));
-
-        Waiting unSavedWaiting = new Waiting(reservation, member, LocalDateTime.now());
-
-        return WaitingResponse.from(waitingRepository.save(unSavedWaiting));
-    }
-
-    public List<WaitingResponse> findWaitings() {
-        List<Waiting> waitings = waitingRepository.findAll();
-
-        return waitings.stream()
-                .map(WaitingResponse::from)
-                .toList();
-    }
-
-    public Waiting findWaitingById(long id) {
-        return waitingRepository.findById(id)
-                .orElseThrow(() -> new RoomEscapeException(WaitingExceptionCode.WAITING_NOT_EXIST_EXCEPTION));
-    }
-
-    public List<WaitingWithRank> findMemberWaitingWithRank(long memberId) {
-        return waitingRepository.findWaitingsWithRankByMemberId(memberId);
-    }
-
-    public long countWaitingRank(WaitingDto waitingDto) {
-        Waiting waiting = waitingRepository.findByReservationDateAndReservationThemeIdAndReservationReservationTimeIdAndMemberId(
-                waitingDto.date(),
-                waitingDto.themeId(),
-                waitingDto.timeId(),
-                waitingDto.memberId()
-        );
-
-        return waitingRepository.countWaitingRankByDateAndThemeIdAndReservationTimeId(
-                waiting.getId(),
-                waitingDto.date(),
-                waitingDto.themeId(),
-                waitingDto.timeId()
-        );
-    }
-
-    public void removeWaiting(long waitingId) {
-        waitingRepository.deleteById(waitingId);
-    }
-
-    private void validateAlreadyReservation(WaitingRequest waitingRequest, long memberId) {
-        boolean existReservation = reservationRepository.existsByDateAndThemeIdAndReservationTimeIdAndMemberId(
-                waitingRequest.date(),
-                waitingRequest.themeId(),
-                waitingRequest.timeId(),
-                memberId
-        );
-
-        if (existReservation) {
-            throw new RoomEscapeException(WaitingExceptionCode.ALREADY_REGISTRATION_EXCEPTION);
-        }
-    }
-
-    private void validateAlreadyWaiting(WaitingRequest waitingRequest, long memberId) {
-        boolean existWaiting = waitingRepository.existsByReservationDateAndReservationThemeIdAndReservationReservationTimeIdAndMemberId(
-                waitingRequest.date(),
-                waitingRequest.themeId(),
-                waitingRequest.timeId(),
-                memberId
-        );
-
-        if (existWaiting) {
-            throw new RoomEscapeException(WaitingExceptionCode.ALREADY_REGISTRATION_EXCEPTION);
-        }
-    }
-}
