+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,0 +1,10 @@
+package roomescape.reservation.dto;
+
+import java.time.LocalDate;
+
+public record ReservationRequest(
+        LocalDate date,
+        long timeId,
+        long themeId
+) {
+}
