+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -1,0 +1,19 @@
+package roomescape.reservation.dto;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import roomescape.reservation.domain.Reservation;
+
+public record ReservationResponse(
+        long id,
+        String memberName,
+        String themeName,
+        LocalDate date,
+        LocalTime startAt
+) {
+    public static ReservationResponse from(Reservation reservation) {
+        return new ReservationResponse(reservation.getId(), reservation.getMember().getName(),
+                reservation.getTheme().getName(), reservation.getDate(),
+                reservation.getReservationTime().getStartAt());
+    }
+}
