+++ src/main/java/roomescape/waiting/domain/WaitingWithRank.java
@@ -1,0 +1,8 @@
+package roomescape.waiting.domain;
+
+public record WaitingWithRank(Waiting waiting, Long rank) {
+
+    public WaitingWithRank(Waiting waiting, int rank) {
+        this(waiting, (long) rank);
+    }
+}
