+++ src/main/java/roomescape/waiting/dto/WaitingRequest.java
@@ -1,0 +1,10 @@
+package roomescape.waiting.dto;
+
+import java.time.LocalDate;
+
+public record WaitingRequest(
+        LocalDate date,
+        long timeId,
+        long themeId
+) {
+}
