+++ src/test/java/roomescape/reservation/controller/ReservationControllerTest.java
@@ -16,15 +16,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.registration.domain.reservation.controller.ReservationController;
 import roomescape.vo.Name;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
 import roomescape.model.ControllerTest;
-import roomescape.registration.domain.reservation.domain.Reservation;
-import roomescape.registration.domain.reservation.dto.ReservationResponse;
-import roomescape.registration.domain.reservation.dto.ReservationTimeAvailabilityResponse;
-import roomescape.registration.domain.reservation.service.ReservationService;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.ReservationTimeAvailabilityResponse;
+import roomescape.reservation.service.ReservationService;
 import roomescape.theme.domain.Theme;
 import roomescape.reservationtime.domain.ReservationTime;
